change type of price in products table

+++ database/migrations/2023_10_03_030400_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('description');
-            $table->float('price');
+            $table->decimal('price', 15, 2);
             $table->string('picture_url');
             $table->integer('user_id');
             $table->timestamps();
